+++ host/modules/api.jsx
@@ -4,5 +4,6 @@ $.global.FishTools = {
     applyEase: _applyEase,
     readVelocity: _readVelocity,
     applyVelocity: _applyVelocity,
+    applyElastic: _applyElastic,
     debugProps: _debugProps
 };